Squashed commit of the following:
commit ee7e8002941d3fd325b5073c83a7f4bac3a6086e
Author: Avinash Madavan <madavan2@illinois.edu>
Date:   Fri Aug 23 20:40:42 2019 -0500

    Update callbacks, bazel build, and preconditioners

commit 59fc9a32255c222dad29ff1211a7ed029c9738c8
Author: Avinash <avinash.madavan@gmail.com>
Date:   Wed Jun 26 12:22:04 2019 -0500

    Bug fixes for PDSS and HDF5 saving.

    HDF5 files were saving integers instead of doubles. PDSS requires boundedness of constraints which was not previously implemented.

commit 4769ef1dbd36d9080d05f2aeb1e6ae669d637492
Author: Avinash <avinash.madavan@gmail.com>
Date:   Sun Jun 23 22:17:57 2019 -0500

    Fix build errors.

    Small errors in naming and fix for errors observed in linux build (AWS CentOS instance).

commit 4fbc39b02c628e8a0f536d49354911d507e2186a
Author: Avinash <avinash.madavan@gmail.com>
Date:   Sun Jun 23 22:16:33 2019 -0500

    Fix Stochastic PD for CVaR.

    There were several bugs in the stochastic primal-dual method that were fixed, including an incorrect projection operation, incorrect resizing due to CVaR, and transposition errors for constraints.

commit 680178fc5c4ff383acdad8593152b20bdeb805ca
Merge: 68ce342 58e4a04
Author: Avinash Madavan <madavan2@illinois.edu>
Date:   Mon Jun 17 14:25:59 2019 -0500

    Merge branch 'develop' into PrimalDualStochastic

commit 68ce342bf13e471530a76836db590ed4c826d77b
Author: Avinash Madavan <madavan2@illinois.edu>
Date:   Wed Jun 12 06:07:16 2019 -0500

    Clean up CMakeLists.txt.

commit fade940c70e555134e0dcfeb513dabdd3c73a018
Author: Avinash Madavan <madavan2@illinois.edu>
Date:   Mon Jun 3 08:16:21 2019 +0545

    Ensure that stochastic program arguments are constant.

    The stochastic program function and constraint evaluation routines should not modify the state or the sample. Add constant keywords to ensure that this is satisfied.

commit 4725b2f672252bda7f59698c037ba9d98f9297ae
Author: Avinash Madavan <madavan2@illinois.edu>
Date:   Mon Jun 3 08:15:04 2019 +0545

    Update callbacks.

    Add method to callbacks that is called on completion. Additionally, update HDF5 save routine to use C API rather than the C++ API. The C++ API has been noted to have issues that are difficult to debug and for convenience and the sake of my sanity, we simply use the C API.

commit 54cefe88cff433c44630eb9af99300e5f7eb033d
Author: Avinash Madavan <madavan2@illinois.edu>
Date:   Fri May 24 22:04:15 2019 -0500

    Update Eigen version.

commit 2ec5bec88eb10ad489f0d93fdf7c1035350274b0
Author: Avinash Madavan <madavan2@illinois.edu>
Date:   Wed May 22 16:52:30 2019 -0500

    Code cleanup

commit 5cac9420c6e558deaf09d5e93c0f57294c76dbaa
Author: Avinash Madavan <madavan2@illinois.edu>
Date:   Mon May 20 00:55:32 2019 -0500

    Code cleanup

commit 8f0ad3969ba8208314e3f52e241dcab180682bdc
Merge: 92b3955 9878976
Author: Avinash Madavan <madavan2@illinois.edu>
Date:   Sat May 18 19:43:02 2019 -0500

    Merge branch 'develop' into PrimalDualStochastic

commit 92b3955e18688ff6bdabc1af5315696207ed3d9a
Author: Avinash Madavan <madavan2@illinois.edu>
Date:   Thu May 16 13:40:50 2019 -0500

    Minor fixes

commit a4442e4464e069da5f00203d491fe97d9f1a2b15
Merge: c630f5d af49008
Author: Avinash Madavan <madavan2@illinois.edu>
Date:   Tue May 14 13:20:00 2019 -0500

    Merge branch 'develop' into PrimalDualStochastic

commit c630f5d9e499e463392431797de34e926c600f63
Merge: 4736786 de61632
Author: Avinash Madavan <madavan2@illinois.edu>
Date:   Tue May 14 01:02:06 2019 -0500

    Merge develop

commit 473678699a28e9260ef82b6200474865de4112f9
Author: Avinash Madavan <madavan2@illinois.edu>
Date:   Sat May 11 14:14:18 2019 -0500

    Add function callback bindings and minor fixes.

commit 3ba53a827115333c9ad283b331d6e825b2741bc4
Merge: 65a7cba bd9b10d
Author: Avinash Madavan <madavan2@illinois.edu>
Date:   Thu May 9 17:20:18 2019 -0500

    Merge branch 'develop' into PrimalDualStochastic

commit 65a7cba5340db600f47854f925b7e1d1bc41891d
Merge: e5db8d2 1ca1848
Author: Avinash Madavan <madavan2@illinois.edu>
Date:   Thu May 9 17:09:05 2019 -0500

    Merge callbacks

commit e5db8d247f18b9140032b4e8fee93e8540dbe46c
Merge: 4384ad1 7b98c3b
Author: Avinash Madavan <madavan2@illinois.edu>
Date:   Thu May 9 16:01:14 2019 -0500

    Merge develop with callbacks

commit 4384ad110179e830c901a63e447ec3f56a7c6997
Author: Avinash Madavan <madavan2@illinois.edu>
Date:   Tue May 7 17:39:09 2019 -0500

    Implement primal-dual stochastic gradient descent method.

    Create an implementation of the stochastic gradient descent method outlined in my soon-to-be-published publication. Additionally, implement the CVaR and expected value cases to take better advantage of the properties of the expected value case. Other minor code cleanup edits made in the python bindings.

+++ include/stuka/dLP/benders.h
@@ -31,7 +31,7 @@ namespace stuka { namespace dLP {
     Eigen::VectorXd x_;                                                 // Current iterate
     size_t n_sub_;
     size_t n_dim_master_;
-    unsigned int n_sub_calls_;
+    size_t n_sub_calls_;
 
     std::unique_ptr<LP::BaseLPSolver> master_solver_;                   // Solver for master problem
     std::vector<BendersSubproblem> subproblems_;                        // Subproblems
